+++ src/main/java/org/patternfly/component/button/Button.java
@@ -198,7 +198,7 @@ public class Button extends BaseComponent<HTMLElement, Button> implements
         if (anchorElement != null) {
             anchorElement.href = href;
         } else {
-            Logger.undefined(componentType(), element(), "Unable to set href: This button is no <a/> button. ");
+            Logger.undefined(componentType(), element(), "Unable to set href: This button is not an <a/> button.");
         }
         return this;
     }
@@ -207,7 +207,7 @@ public class Button extends BaseComponent<HTMLElement, Button> implements
         if (anchorElement != null) {
             anchorElement.target = target;
         } else {
-            Logger.undefined(componentType(), element(), "Unable to set target: This button is no <a/> button. ");
+            Logger.undefined(componentType(), element(), "Unable to set target: This button is not an <a/> button.");
         }
         return this;
     }