+++ src/main/java/org/patternfly/component/list/SimpleListSubComponent.java
@@ -21,9 +21,9 @@ import org.patternfly.component.SubComponent;
 
 import elemental2.dom.HTMLElement;
 
-abstract class ListSubComponent<E extends HTMLElement, B extends TypedBuilder<E, B>> extends SubComponent<E, B> {
+abstract class SimpleListSubComponent<E extends HTMLElement, B extends TypedBuilder<E, B>> extends SubComponent<E, B> {
 
-    ListSubComponent(String name, E element) {
-        super(ComponentType.List, name, element);
+    SimpleListSubComponent(String name, E element) {
+        super(ComponentType.SimpleList, name, element);
     }
 }