+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Progress component (#81)
+- Simple list component (#84)
 
 ## [0.1.4] - 2024-01-25